English is not special. Also, speed up RDF import.
When importing the data from the RDF, anywhere the language
is omitted, assume it's English and set that language
explicitly on the imported data. Once this is done, we
should not need to implement any default language for
the license metadata, it'll all have languages.

While testing this, I ended up speeding up the RDF
import. It was taking about 5 minutes on my laptop. Now
it's 7 seconds, fast enough to just include in a migration
so the tests have real data to work on.

Also wrote unit tests to check that the imported data is what
we expect.